Introduce JSON queries in the WebUI.
We create etorrent_query as a new module which is used for querying
the internal state of the etorrent system. Next, we build a wrapper
in the etorrent_webui module, outputting JSON to the front-end upon
query.

While here, use the new etorrent_query module as a factoring module
for the original queries.