Properly consider invalid pieces in check_torrent(..)
We will have to check all pieces when we are rechecking a
torrent. Thus, we create the whole set of all pieces and proceed to
filter it. We select all the pieces which are not valid and call those
the Invalid pieces. Then we check if this list of pieces is empty or
not.